refactor(ui): fold the muted-foreground guard into one text-tone contract
muted-foreground-contrast.test.ts already parsed tokens.css and implemented
OKLCH -> contrast to assert the AA floors from MUL-5447. Keeping a second
file doing the same thing for --faint-foreground would be two guards that
drift apart, so this merges them and keeps the older file's better helpers
(var() alias resolution, 8-bit quantisation).

The merge also caught a real hole: .landing-light in custom.css re-declares
the whole light palette so landing components stay light under next-themes'
.dark class. It had no --faint-foreground, so the landing chevrons would
have inherited the .dark value while sitting on a white surface. Added,
and the sync assertion now covers both tones instead of just muted.
