+++ apps/web/app/text-contrast.test.ts
@@ -3,111 +3,161 @@ import { resolve, join, relative } from "node:path";
 import { describe, expect, it } from "vitest";
 
 /**
- * Can every piece of product text actually be read, and is hierarchy still
- * expressed with tokens rather than with alpha?
+ * Is every tone the product paints text and marks with actually legible, and is
+ * hierarchy still expressed with tokens rather than with alpha?
  *
- * The two questions belong in one guard because the answer to the first used to
+ * The two questions are one contract, because the answer to the first used to
  * depend on the second. `text-muted-foreground/70` reads as "muted, a bit
- * quieter"; on a light surface it is 2.69:1, which is not quiet, it is gone.
- * 152 call sites spread across /30 to /80 and every one of them failed AA in
- * light mode, because alpha was the only tool available for "weaker than
- * muted" — the palette had no step below it.
+ * quieter"; on a light surface it is 2.69:1, which is not quieter, it is gone.
+ * 152 call sites spread across /30 to /80 and every one failed AA in light
+ * mode — not because anyone chose an illegible grey, but because the palette
+ * stopped at --muted-foreground and alpha was the only way to say "quieter".
  *
- * So the palette got the step, and it is deliberately a NON-TEXT step:
- * --faint-foreground clears 3:1 (WCAG 1.4.11, for icons and glyphs) but not
- * 4.5:1, because there is no room for a third readable text tone. AA caps a
- * lighter text tone at L 0.523 on our darkest light surface and muted already
- * sits at 0.505 — a tier 0.018 apart is a tier nobody can see. Text has one
+ * So the palette got one more step, and it is deliberately a NON-TEXT step.
+ * --faint-foreground clears 3:1 (WCAG 1.4.11, icons and glyphs) but not 4.5:1,
+ * because there is no room for a third readable text tone: AA caps a lighter
+ * text tone at L 0.523 on our darkest light surface and muted already sits at
+ * 0.505. A tier 0.018 apart is a tier nobody can see, so text keeps exactly one
  * floor and it is --muted-foreground.
  *
- * The contrast assertions recompute from tokens.css rather than hard-coding
- * ratios, so editing a token value is what fails the test — which is the only
- * way a colour guard stays true. Ratios are WCAG 2.x relative luminance.
+ * Assertions recompute contrast from the token files rather than hard-coding
+ * ratios, so editing a token value is what fails the test — the only way a
+ * colour guard stays true. Ratios are WCAG 2.x relative luminance.
  */
 
 const repoRoot = resolve(process.cwd(), "../..");
-const tokensPath = resolve(repoRoot, "packages/ui/styles/tokens.css");
 
-// ── colour maths ───────────────────────────────────────────────────────────
+/** WCAG 1.4.3 for body and label text; 1.4.11 for icons and other marks. */
+const WCAG_AA_NORMAL_TEXT = 4.5;
+const WCAG_AA_NON_TEXT = 3.0;
+
+/**
+ * Surfaces that product text and marks are painted on. Border/input/ring tokens
+ * are excluded: they are not text backgrounds, and several carry alpha, which
+ * has no single contrast answer without knowing what is underneath.
+ */
+const backgroundTokens = [
+  "--app-shell",
+  "--page-canvas",
+  "--background",
+  "--surface",
+  "--surface-raised",
+  "--surface-hover",
+  "--surface-selected",
+  "--card",
+  "--popover",
+  "--muted",
+  "--secondary",
+  "--accent",
+  "--sidebar",
+  "--sidebar-accent",
+];
 
 type Rgb = [number, number, number];
 
-/** OKLCH -> linear sRGB -> gamma-encoded sRGB, clamped to gamut. */
-function oklchToSrgb(l: number, c: number, hDeg: number): Rgb {
-  const h = (hDeg * Math.PI) / 180;
-  const a = c * Math.cos(h);
-  const b = c * Math.sin(h);
-  const lc = (l + 0.3963377774 * a + 0.2158037573 * b) ** 3;
-  const mc = (l - 0.1055613458 * a - 0.0638541728 * b) ** 3;
-  const sc = (l - 0.0894841775 * a - 1.291485548 * b) ** 3;
-  const linear = [
-    4.0767416621 * lc - 3.3077115913 * mc + 0.2309699292 * sc,
-    -1.2684380046 * lc + 2.6097574011 * mc - 0.3413193965 * sc,
-    -0.0041960863 * lc - 0.7034186147 * mc + 1.707614701 * sc,
+function readBlock(source: string, selector: string): Map<string, string> {
+  const start = source.indexOf(`${selector} {`);
+  if (start < 0) throw new Error(`${selector} block not found`);
+  const end = source.indexOf("\n}", start);
+  if (end < 0) throw new Error(`${selector} block is unterminated`);
+
+  const declarations = new Map<string, string>();
+  for (const match of source.slice(start, end).matchAll(/(--[\w-]+):\s*([^;]+);/g)) {
+    const [, name, value] = match;
+    if (name && value) declarations.set(name, value.trim());
+  }
+  return declarations;
+}
+
+/** Follows `--card: var(--surface)` style indirection to a literal colour. */
+function resolveToken(declarations: Map<string, string>, name: string): string {
+  let current = name;
+
+  for (let hops = 0; hops < 8; hops++) {
+    const value = declarations.get(current);
+    if (value === undefined) {
+      throw new Error(`${current} is not declared (while resolving ${name})`);
+    }
+    const alias = /^var\((--[\w-]+)\)$/.exec(value)?.[1];
+    if (!alias) return value;
+    current = alias;
+  }
+  throw new Error(`${name} did not resolve to a literal colour`);
+}
+
+const clamp01 = (x: number) => Math.min(1, Math.max(0, x));
+const encodeSrgb = (c: number) =>
+  c <= 0.0031308 ? 12.92 * c : 1.055 * Math.pow(c, 1 / 2.4) - 0.055;
+const decodeSrgb = (c: number) =>
+  c <= 0.04045 ? c / 12.92 : Math.pow((c + 0.055) / 1.055, 2.4);
+
+/**
+ * Quantises to 8-bit on purpose: contrast is judged on what the display
+ * actually paints, not on the unrounded float behind it.
+ */
+function oklchToRgb(value: string): Rgb {
+  const match = /^oklch\(\s*([\d.]+)\s+([\d.]+)\s+([\d.]+)\s*\)$/.exec(value);
+  if (!match?.[1] || !match[2] || !match[3]) {
+    throw new Error(`expected an alpha-free oklch() colour, got "${value}"`);
+  }
+
+  const lightness = Number(match[1]);
+  const chroma = Number(match[2]);
+  const hue = (Number(match[3]) * Math.PI) / 180;
+  const a = chroma * Math.cos(hue);
+  const b = chroma * Math.sin(hue);
+
+  const l = (lightness + 0.3963377774 * a + 0.2158037573 * b) ** 3;
+  const m = (lightness - 0.1055613458 * a - 0.0638541728 * b) ** 3;
+  const s = (lightness - 0.0894841775 * a - 1.291485548 * b) ** 3;
+
+  const channel = (linear: number) =>
+    Math.round(clamp01(encodeSrgb(clamp01(linear))) * 255);
+
+  return [
+    channel(4.0767416621 * l - 3.3077115913 * m + 0.2309699292 * s),
+    channel(-1.2684380046 * l + 2.6097574011 * m - 0.3413193965 * s),
+    channel(-0.0041960863 * l - 0.7034186147 * m + 1.707614701 * s),
   ];
-  return linear.map((v) => {
-    const x = Math.min(1, Math.max(0, v));
-    return x <= 0.0031308 ? 12.92 * x : 1.055 * x ** (1 / 2.4) - 0.055;
-  }) as Rgb;
 }
 
 function relativeLuminance([r, g, b]: Rgb): number {
-  const lin = (v: number) => (v <= 0.04045 ? v / 12.92 : ((v + 0.055) / 1.055) ** 2.4);
-  return 0.2126 * lin(r) + 0.7152 * lin(g) + 0.0722 * lin(b);
-}
-
-function contrastRatio(fg: Rgb, bg: Rgb): number {
-  const [hi, lo] = [relativeLuminance(fg), relativeLuminance(bg)].sort((x, y) => y - x);
-  return (hi! + 0.05) / (lo! + 0.05);
-}
-
-// ── token parsing ──────────────────────────────────────────────────────────
-
-const tokens = readFileSync(tokensPath, "utf8");
-
-/**
- * Pull one theme block's own declarations. `:root` also contains the `@theme`
- * blocks' siblings, so the match is anchored on the selector and stops at the
- * first closing brace at column 0.
- */
-function themeBlock(selector: string): string {
-  const start = tokens.indexOf(`${selector} {`);
-  if (start === -1) throw new Error(`no ${selector} block in tokens.css`);
-  const end = tokens.indexOf("\n}", start);
-  return tokens.slice(start, end);
-}
-
-function readOklch(block: string, name: string): Rgb {
-  const match = block.match(
-    new RegExp(String.raw`--${name}:\s*oklch\(([\d.]+)\s+([\d.]+)\s+([\d.]+)\)`),
+  return (
+    0.2126 * decodeSrgb(r / 255) +
+    0.7152 * decodeSrgb(g / 255) +
+    0.0722 * decodeSrgb(b / 255)
   );
-  if (!match) throw new Error(`--${name} is not a literal oklch() in this theme block`);
-  return oklchToSrgb(Number(match[1]), Number(match[2]), Number(match[3]));
 }
 
-/**
- * Every surface product text sits on. Aliases are omitted: --background is
- * --page-canvas, --card and --popover are --surface, and --accent/--secondary
- * share --muted's value, so covering the bases covers them.
- */
-const SURFACES = [
-  "surface",
-  "page-canvas",
-  "app-shell",
-  "muted",
-  "surface-selected",
-  "sidebar",
-  "sidebar-accent",
-] as const;
+function contrastRatio(foreground: Rgb, background: Rgb): number {
+  const a = relativeLuminance(foreground);
+  const b = relativeLuminance(background);
+  const [lighter, darker] = a > b ? [a, b] : [b, a];
+  return (lighter + 0.05) / (darker + 0.05);
+}
 
-const THEMES = [
-  ["light", ":root"],
-  ["dark", ".dark"],
-] as const;
+function expectTonePasses(
+  declarations: Map<string, string>,
+  tone: string,
+  floor: number,
+) {
+  const foreground = oklchToRgb(resolveToken(declarations, tone));
 
-/** WCAG 1.4.3 for body and label text; 1.4.11 for icons and other marks. */
-const AA_TEXT = 4.5;
-const AA_NON_TEXT = 3.0;
+  for (const token of backgroundTokens) {
+    const background = oklchToRgb(resolveToken(declarations, token));
+    const ratio = contrastRatio(foreground, background);
+
+    expect(
+      Number(ratio.toFixed(2)),
+      `${tone} on ${token} is ${ratio.toFixed(2)}:1`,
+    ).toBeGreaterThanOrEqual(floor);
+  }
+}
+
+const tokensCss = () =>
+  readFileSync(resolve(repoRoot, "packages/ui/styles/tokens.css"), "utf8");
+const landingCss = () =>
+  readFileSync(resolve(repoRoot, "apps/web/app/custom.css"), "utf8");
 
 // ── source scanning ────────────────────────────────────────────────────────
 
@@ -121,12 +171,18 @@ const sourceExtensions = [".ts", ".tsx", ".css"];
  * on purpose: they paint text on photos, gradients, and inverted cards, where
  * no solid secondary token exists and alpha is the honest mechanism.
  */
-const guardedColors = ["foreground", "muted-foreground", "sidebar-foreground", "destructive", "current"];
+const guardedColors = [
+  "foreground",
+  "muted-foreground",
+  "sidebar-foreground",
+  "destructive",
+  "current",
+];
 
 /**
  * Alpha behind an interaction variant is transient feedback, not a hierarchy
- * level — the resting state is what carries the contrast obligation, and it is
- * solid. `dark:` is not on this list: it paints a resting state.
+ * level — the resting state carries the contrast obligation, and it is solid.
+ * `dark:` is deliberately absent: it paints a resting state.
  */
 const stateVariants =
   /\b(?:hover|focus|focus-visible|focus-within|active|disabled|visited|group-hover|group-focus|peer-hover|peer-focus|aria-disabled|data-disabled|data-\[disabled)\b/;
@@ -159,32 +215,57 @@ function collectSourceFiles(dir: string, found: string[] = []): string[] {
   return found;
 }
 
-// ── the guard ──────────────────────────────────────────────────────────────
+// ── the contract ───────────────────────────────────────────────────────────
 
 describe("text contrast", () => {
-  describe.each(THEMES)("%s theme", (_theme, selector) => {
-    const block = themeBlock(selector);
-    const surfaces = SURFACES.map((name) => [name, readOklch(block, name)] as const);
-
-    it.each(surfaces)("muted-foreground clears AA text contrast on --%s", (_name, bg) => {
-      expect(contrastRatio(readOklch(block, "muted-foreground"), bg)).toBeGreaterThanOrEqual(
-        AA_TEXT,
-      );
+  describe("--muted-foreground, the floor for text", () => {
+    it("clears WCAG AA on every light surface", () => {
+      expectTonePasses(readBlock(tokensCss(), ":root"), "--muted-foreground", WCAG_AA_NORMAL_TEXT);
     });
 
-    it.each(surfaces)("faint-foreground clears non-text contrast on --%s", (_name, bg) => {
-      expect(contrastRatio(readOklch(block, "faint-foreground"), bg)).toBeGreaterThanOrEqual(
-        AA_NON_TEXT,
-      );
-    });
-
-    it("faint-foreground is quieter than muted-foreground, not louder", () => {
-      const surface = readOklch(block, "surface");
-      expect(contrastRatio(readOklch(block, "faint-foreground"), surface)).toBeLessThan(
-        contrastRatio(readOklch(block, "muted-foreground"), surface),
-      );
+    it("clears WCAG AA on every dark surface", () => {
+      expectTonePasses(readBlock(tokensCss(), ".dark"), "--muted-foreground", WCAG_AA_NORMAL_TEXT);
     });
   });
+
+  describe("--faint-foreground, for marks that are not text", () => {
+    it("clears non-text contrast on every light surface", () => {
+      expectTonePasses(readBlock(tokensCss(), ":root"), "--faint-foreground", WCAG_AA_NON_TEXT);
+    });
+
+    it("clears non-text contrast on every dark surface", () => {
+      expectTonePasses(readBlock(tokensCss(), ".dark"), "--faint-foreground", WCAG_AA_NON_TEXT);
+    });
+
+    it.each([
+      ["light", ":root"],
+      ["dark", ".dark"],
+    ])("stays quieter than muted-foreground in %s mode", (_mode, selector) => {
+      const declarations = readBlock(tokensCss(), selector);
+      const surface = oklchToRgb(resolveToken(declarations, "--surface"));
+      const faint = oklchToRgb(resolveToken(declarations, "--faint-foreground"));
+      const muted = oklchToRgb(resolveToken(declarations, "--muted-foreground"));
+
+      expect(
+        contrastRatio(faint, surface),
+        "faint must read as the quieter step, or the two tiers have swapped",
+      ).toBeLessThan(contrastRatio(muted, surface));
+    });
+  });
+
+  // The landing route tree re-declares the light palette so token-driven
+  // components stay light under next-themes' `.dark` class. That copy is only
+  // correct while it matches the source, so drift here is a bug in itself — a
+  // tone missing from the copy silently inherits its `.dark` value on a white
+  // surface.
+  it.each(["--muted-foreground", "--faint-foreground"])(
+    "keeps the landing-light copy of %s in sync with the light token",
+    (tone) => {
+      expect(resolveToken(readBlock(landingCss(), ".landing-light"), tone)).toBe(
+        resolveToken(readBlock(tokensCss(), ":root"), tone),
+      );
+    },
+  );
 
   it("product UI expresses text hierarchy with tokens, not alpha", () => {
     const violations: string[] = [];
